search: move zoekt global job to zoekt package
