+++ internal/search/job/printers.go
@@ -12,7 +12,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/search/searcher"
 	"github.com/sourcegraph/sourcegraph/internal/search/structural"
 	"github.com/sourcegraph/sourcegraph/internal/search/symbol"
-	"github.com/sourcegraph/sourcegraph/internal/search/textsearch"
 	"github.com/sourcegraph/sourcegraph/internal/search/zoekt"
 )
 
@@ -44,7 +43,7 @@ func SexpFormat(job Job, sep, indent string) string {
 			*zoekt.ZoektRepoSubsetSearch,
 			*searcher.Searcher,
 			*run.RepoSearch,
-			*textsearch.RepoUniverseTextSearch,
+			*zoekt.GlobalSearch,
 			*structural.StructuralSearch,
 			*commit.CommitSearch,
 			*symbol.RepoSubsetSymbolSearch,
@@ -213,7 +212,7 @@ func PrettyMermaid(job Job) string {
 			*zoekt.ZoektRepoSubsetSearch,
 			*searcher.Searcher,
 			*run.RepoSearch,
-			*textsearch.RepoUniverseTextSearch,
+			*zoekt.GlobalSearch,
 			*structural.StructuralSearch,
 			*commit.CommitSearch,
 			*symbol.RepoSubsetSymbolSearch,
@@ -338,7 +337,7 @@ func toJSON(job Job, verbose bool) interface{} {
 			*zoekt.ZoektRepoSubsetSearch,
 			*searcher.Searcher,
 			*run.RepoSearch,
-			*textsearch.RepoUniverseTextSearch,
+			*zoekt.GlobalSearch,
 			*structural.StructuralSearch,
 			*commit.CommitSearch,
 			*symbol.RepoSubsetSymbolSearch,
